% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_chilling_units.R
\name{calculate_chilling_units}
\alias{calculate_chilling_units}
\title{Calculate Chilling Units}
\usage{
calculate_chilling_units(
  data,
  tmin_col = "tmin",
  tmax_col = "tmax",
  tmin_chill = 0,
  tmax_chill = 7.2
)
}
\arguments{
\item{data}{A dataframe with climate data (output of import_climate_data)}

\item{tmin_col}{Name of the minimum temperature column (default: "tmin")}

\item{tmax_col}{Name of the maximum temperature column (default: "tmax")}

\item{tmin_chill}{Minimum temperature for chilling (default: 0)}

\item{tmax_chill}{Maximum temperature for chilling (default: 7.2)}
}
\value{
A dataframe with daily and cumulative chilling units
}
\description{
Calculate Chilling Units
}
\examples{
df <- import_climate_data("climate.csv")
cu <- calculate_chilling_units(df)
}
