% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_sample}
\alias{climate_sample}
\title{Données climatiques d'exemple}
\format{
Un dataframe avec 365 lignes et 3 variables :
\describe{
\item{date}{Date au format Date}
\item{tmin}{Température minimale journalière en degrés Celsius}
\item{tmax}{Température maximale journalière en degrés Celsius}
}
}
\source{
Données simulées
}
\usage{
climate_sample
}
\description{
Jeu de données climatiques journalières simulées pour l'année 2023,
utilisé pour illustrer les fonctions du package olivePhenR.
}
\keyword{datasets}
