% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_climate_rasters.R
\name{load_climate_rasters}
\alias{load_climate_rasters}
\title{Load Climate Rasters}
\usage{
load_climate_rasters(path, variable = "tmin", extent = NULL)
}
\arguments{
\item{path}{Path to the folder containing raster files}

\item{variable}{Climate variable to load: "tmin" or "tmax" (default: "tmin")}

\item{extent}{Optional extent to crop the raster (default: NULL)}
}
\value{
A SpatRaster object
}
\description{
Load Climate Rasters
}
\examples{
r <- load_climate_rasters(path = "data/rasters", variable = "tmin")
}
