% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sentinel_data.R
\name{load_sentinel_data}
\alias{load_sentinel_data}
\title{Load Sentinel-2 Data via Copernicus STAC}
\usage{
load_sentinel_data(
  start_date,
  end_date,
  extent = c(-13, 27, -1, 36),
  max_cloud = 20
)
}
\arguments{
\item{start_date}{Start date for data search (format: "YYYY-MM-DD")}

\item{end_date}{End date for data search (format: "YYYY-MM-DD")}

\item{extent}{Vector c(xmin, ymin, xmax, ymax) for the study area}

\item{max_cloud}{Maximum cloud cover percentage (default: 20)}
}
\value{
A dataframe with available Sentinel-2 scenes information
}
\description{
Load Sentinel-2 Data via Copernicus STAC
}
\examples{
\dontrun{
extent_maroc <- c(-13, 27, -1, 36)
scenes <- load_sentinel_data(
  start_date = "2023-03-01",
  end_date   = "2023-04-30",
  extent     = extent_maroc
)
}
}
