% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_flowering_date.R
\name{map_flowering_date}
\alias{map_flowering_date}
\title{Map Flowering Date}
\usage{
map_flowering_date(
  tmin_raster,
  tmax_raster,
  tbase = 10,
  threshold_flowering = 300
)
}
\arguments{
\item{tmin_raster}{SpatRaster of minimum temperatures}

\item{tmax_raster}{SpatRaster of maximum temperatures}

\item{tbase}{Base temperature for GDD calculation (default: 10)}

\item{threshold_flowering}{GDD threshold for flowering (default: 300)}
}
\value{
A SpatRaster with estimated flowering dates (day of year)
}
\description{
Map Flowering Date
}
\examples{
r_tmin <- load_climate_rasters(path = "data/rasters", variable = "tmin")
r_tmax <- load_climate_rasters(path = "data/rasters", variable = "tmax")
flowering_map <- map_flowering_date(r_tmin, r_tmax)
}
