% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_frost_risk.R
\name{map_frost_risk}
\alias{map_frost_risk}
\title{Map Frost Risk}
\usage{
map_frost_risk(tmin_raster, flowering_raster, frost_threshold = 0)
}
\arguments{
\item{tmin_raster}{SpatRaster of minimum temperatures}

\item{flowering_raster}{SpatRaster of flowering dates (output of map_flowering_date)}

\item{frost_threshold}{Temperature threshold for frost (default: 0)}
}
\value{
A SpatRaster with frost risk index
}
\description{
Map Frost Risk
}
\examples{
risk_map <- map_frost_risk(r_tmin, flowering_map)
}
