% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_risk_maps.R
\name{plot_risk_maps}
\alias{plot_risk_maps}
\title{Plot Risk Maps}
\usage{
plot_risk_maps(raster, title = "Carte phenologique", export_path = NULL)
}
\arguments{
\item{raster}{A SpatRaster to plot}

\item{title}{Title of the map (default: "Carte phenologique")}

\item{export_path}{Optional path to export PNG/PDF (default: NULL)}
}
\value{
A plot
}
\description{
Plot Risk Maps
}
\examples{
plot_risk_maps(flowering_map, title = "Date de floraison")
}
